Moved CobaltMetricsServiceClient to exposed class
