+++ cobalt/browser/BUILD.gn
@@ -28,8 +28,8 @@ source_set("browser") {
     "cobalt_single_render_process_observer.h",
     "cobalt_web_contents_observer.cc",
     "cobalt_web_contents_observer.h",
-    "metrics/metrics_service_wrapper.cc",
-    "metrics/metrics_service_wrapper.h",
+    "metrics/cobalt_metrics_service_client.cc",
+    "metrics/cobalt_metrics_service_client.h",
   ]
 
   if (is_androidtv) {
@@ -51,6 +51,8 @@ source_set("browser") {
     "//components/js_injection/browser:browser",
     "//components/prefs",
     "//components/metrics",
+    "//components/variations",
+    "//components/metrics:test_support",
     "//content/public/browser",
     "//content/shell:content_shell_lib",
   ]
