+++ cobalt/browser/cobalt_content_browser_client.cc
@@ -18,6 +18,7 @@
 
 #include "cobalt/browser/cobalt_browser_interface_binders.h"
 #include "cobalt/browser/cobalt_web_contents_observer.h"
+#include "cobalt/browser/metrics/cobalt_metrics_service_client.h"
 #include "cobalt/media/service/mojom/video_geometry_setter.mojom.h"
 #include "cobalt/user_agent/user_agent_platform_info.h"
 #include "content/public/browser/browser_thread.h"
@@ -50,6 +51,9 @@ class CobaltBrowserMainParts : public content::ShellBrowserMainParts {
   int PreCreateThreads() override {
     // TODO(b/372559349): setup metrics similarly to what SetupMetrics() does
     // when called from ChromeBrowserMainParts::PreCreateThreadsImpl().
+
+    //CobaltMetricsServiceClient le_metrics;
+
     return ShellBrowserMainParts::PreCreateThreads();
   }
